refactor(p2p): some improvements from reviewers
