+++ daemons_runner.py
@@ -4,13 +4,13 @@ import asyncio
 
 from daemons.data_collector import DataCollector
 
-async def data_collector():
+async def data_collector() -> None:
     data_collector = DataCollector()
     while True:
         asyncio.create_task(data_collector.run())
         await asyncio.sleep(1)
 
-if __name__ == "__main__":
+def main() -> None:
     loop = asyncio.get_event_loop()
     try:
         asyncio.ensure_future(data_collector())
@@ -20,3 +20,6 @@ if __name__ == "__main__":
     finally:
         print("Closing Loop")
         loop.close()
+
+if __name__ == "__main__":
+    main()
